requirements.txt
discord.py
python-dotenv
pymongo
pillow
requests